chore: fix go linting \n Signed-off-by: sk31337 <sk31337@users.noreply.github.com>
